requirements.txt
absl-py==1.4.0
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.2.2
anyio==3.6.2
APScheduler==3.6.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.6.0
asttokens==2.0.8
astunparse==1.6.3
async-generator==1.10
async-timeout==4.0.2
attrs==22.1.0
Babel==2.11.0
backcall==0.2.0
beautifulsoup4==4.11.1
beautifultable==1.1.0
bleach==5.0.1
blinker==1.6.2
boto3==1.26.89
botocore==1.29.89
cachetools==4.2.2
certifi==2022.6.15
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
click==8.0.4
click-plugins==1.1.1
cligj==0.7.2
climage==0.1.3
colorama==0.4.6
crypto==1.4.1
cryptography==39.0.2
cssselect==1.2.0
cssutils==2.6.0
customtkinter==5.1.2
cycler==0.11.0
Cython==0.29.33
darkdetect==0.8.0
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
discord==2.2.2
discord.py==2.2.2
distlib==0.3.6
Django==4.2.1
django-upgrade==1.13.0
entrypoints==0.4
exceptiongroup==1.1.0
executing==1.1.1
fastapi==0.95.2
fastjsonschema==2.16.2
ffmpy==0.3.0
filelock==3.12.0
filterpy==1.4.5
Fiona==1.9.2
fire==0.5.0
Flask==2.2.2
flatbuffers==23.3.3
fonttools==4.37.1
frozenlist==1.3.3
fsspec==2023.5.0
gast==0.4.0
geopandas==0.12.2
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.16.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googletrans==3.0.0
gradio==3.32.0
gradio_client==0.2.5
grpcio==1.51.3
gTTS==2.3.1
h11==0.9.0
h2==3.2.0
h5py==3.8.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.13.1
hyperframe==5.2.0
idna==2.10
imageai==3.0.3
imageio==2.26.1
importlib-metadata==6.6.0
iniconfig==2.0.0
ipykernel==6.16.2
ipython==8.5.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
iso3166==2.1.1
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
json5==0.9.10
jsonschema==4.16.0
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-server==1.21.0
jupyter_client==7.4.4
jupyter_core==4.11.2
jupyterlab==3.5.1
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
jupyterlab_server==2.16.3
kdtree==0.16
keras==2.11.0
kiwisolver==1.4.4
lab==7.2
lazy_loader==0.1
libclang==15.0.6.1
linkify-it-py==2.0.2
lxml==4.9.2
Markdown==3.4.1
markdown-it-py==2.2.0
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.3
mdurl==0.1.2
mistune==2.0.4
mock==5.0.1
MouseInfo==0.1.3
mpmath==1.3.0
multidict==6.0.4
munch==2.5.0
Naked==0.1.32
nbclassic==0.4.5
nbclient==0.7.0
nbconvert==7.2.2
nbformat==5.7.0
nest-asyncio==1.5.6
networkx==3.0
neuralintents==0.0.4
nltk==3.7
notebook==6.5.1
notebook_shim==0.2.0
numpy==1.23.2
oauthlib==3.2.2
openai==0.26.5
opencv-python==4.7.0.72
opt-einsum==3.3.0
orjson==3.8.14
outcome==1.2.0
packaging==21.3
pandas==1.4.4
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.5.0
playsound==1.3.0
plotly==5.10.0
pluggy==1.0.0
premailer==3.10.0
prettytable==3.6.0
prometheus-client==0.15.0
prompt-toolkit==3.0.31
protobuf==3.19.6
psutil==5.9.3
psycopg2==2.9.3
pure-eval==0.2.2
pyarrow==12.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.53
pybboxes==0.1.6
pycparser==2.21
pydantic==1.10.8
pydeck==0.8.1b0
pydub==0.25.1
pyfiglet==0.8.post1
pygame==2.2.0
PyGetWindow==0.0.9
Pygments==2.13.0
Pympler==1.0.1
PyMsgBox==1.0.9
pyOpenSSL==23.0.0
pyparsing==3.0.9
pyperclip==1.8.2
pyproj==3.5.0
PyRect==0.2.0
pyrsistent==0.18.1
PyScreeze==0.1.28
PySocks==1.7.1
pytesseract==0.3.10
pytest==7.3.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
python-telegram-bot==13.15
pytube==12.1.3
pytweening==1.0.4
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
pywhatkit==5.4
pywin32==304
pywinpty==2.0.8
PyYAML==6.0
pyzmq==24.0.1
qtconsole==5.3.2
QtPy==2.2.1
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
rfc3986==1.5.0
rich==13.3.5
rsa==4.9
s3transfer==0.6.0
sahi==0.11.12
scikit-image==0.20.0
scikit-learn==1.1.2
scipy==1.9.3
seaborn==0.12.1
selenium==4.1.3
semantic-version==2.10.0
Send2Trash==1.8.0
sentry-sdk==1.17.0
shapely==2.0.1
shellescape==3.8.1
SimpleITK==2.2.1
simplejson==3.18.0
six==1.16.0
sklearn==0.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
stack-data==0.5.1
starlette==0.27.0
streamlit==1.22.0
sympy==1.12
tenacity==8.1.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.1
tensorflow-estimator==2.11.0
tensorflow-intel==2.11.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
terminado==0.17.0
terminaltables==3.1.10
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tifffile==2023.3.15
tinycss2==1.2.1
tokenize-rt==5.0.0
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==1.13.1
torchvision==0.14.1
tornado==6.3.1
tqdm==4.64.1
traitlets==5.5.0
trio==0.22.0
trio-websocket==0.9.2
ttkbootstrap==1.10.1
txt2tags==3.7
typing_extensions==4.5.0
tzdata==2022.2
tzlocal==4.2
uc-micro-py==1.0.2
ultralytics==8.0.55
urllib3==1.26.12
urllib3-secure-extra==0.1.0
uvicorn==0.22.0
validators==0.20.0
virtualenv==20.23.0
watchdog==3.0.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.1
websockets==11.0.3
Werkzeug==2.2.2
widgetsnbextension==4.0.3
wikipedia==1.4.0
wrapt==1.15.0
wsproto==1.2.0
yagmail==0.15.293
yarl==1.8.2
yolov5==7.0.9
zipp==3.15.0
